ユーザーの退会処理をretirement#createへ移行
